Adopt WXT.
Fixes broken build.
