+++ .gitignore
@@ -25,3 +25,5 @@ private
 .aider*
 
 dist.zip
+
+.wxt
