+++ packages/extension/eslint.config.mjs
@@ -1,0 +1,7 @@
+import rootConfig from '../../eslint.config.mjs';
+import autoImports from './.wxt/eslint-auto-imports.mjs';
+
+export default [
+  ...rootConfig,
+  autoImports,
+];
