+++ packages/extension/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "extension",
+  "name": "Empower Poster",
   "private": true,
-  "description": "Browser extension that posts Empower data to a user-provided endpoint",
-  "version": "0.2.5",
+  "description": "Empower Poster Extension",
+  "version": "0.2.6",
   "author": "Ted Romer <tromer@gmail.com>",
   "license": "MIT",
   "keywords": [
@@ -18,37 +18,33 @@
     "url": "https://github.com/righteffort/empower-poster/issues"
   },
   "type": "module",
-  "engines": {
-    "node": ">=22"
-  },
-  "packageManager": "yarn@4.12.0",
   "scripts": {
-    "build": "vite build",
-    "build:check": "yarn check && yarn build",
-    "dist": "rm -f dist.zip && yarn build && cd dist && zip -r ../dist.zip *",
-    "cli": "tsc -p tsconfig-cli.json && node dist-cli/cli/cli.js",
-    "clean": "rm -rf dist dist-cli",
+    "dev": "wxt",
+    "dev:firefox": "wxt -b firefox",
+    "build": "wxt build",
+    "build:firefox": "wxt build -b firefox",
+    "zip": "wxt zip",
+    "zip:firefox": "wxt zip -b firefox",
+    "clean": "rm -rf dist",
     "typecheck": "tsc --noEmit",
-    "lint": "eslint '{src,tests}/**/*.{js,ts}'",
+    "lint": "eslint '{.,src,tests}/**/*.{js,ts}'",
     "lint:fix": "yarn lint --fix",
-    "format:check": "prettier '{src,tests}/**/*.{js,ts}' --check",
-    "format": "prettier '{src,tests}/**/*.{js,ts}' --write",
+    "format:check": "prettier '*{.,src,tests}/**/*.{js,ts}' --check",
+    "format": "prettier '{.,src,tests}/**/*.{js,ts}' --write",
     "check": "yarn typecheck && yarn lint && yarn format:check",
     "fix": "yarn lint:fix && yarn format",
     "test:watch": "vitest",
     "test": "vitest run",
-    "ci": "yarn check && yarn test"
+    "ci": "yarn check && yarn test",
+    "postinstall": "wxt prepare"
   },
   "devDependencies": {
-    "@eslint/js": "^9.35.0",
-    "@types/node": "^25.0.3",
-    "@types/webextension-polyfill": "^0.10.0",
-    "eslint": "^9.35.0",
-    "prettier": "^3.8.1",
-    "typescript": "^5.9.2",
-    "typescript-eslint": "^8.43.0",
-    "vite": "^8.0.16",
-    "vitest": "^4.0.16",
-    "webextension-polyfill": "^0.10.0"
+    "@eslint/js": "^10.0.1",
+    "eslint": "^10.4.1",
+    "prettier": "^3.8.3",
+    "typescript": "^5.9.3",
+    "typescript-eslint": "^8.60.1",
+    "vitest": "^4.1.8",
+    "wxt": "^0.20.26"
   }
 }
