+++ packages/extension/src/entrypoints/content/processing.ts
