+++ packages/extension/src/entrypoints/onboarding/index.html
