+++ packages/extension/src/entrypoints/onboarding/onboarding.ts
@@ -1,6 +1,6 @@
 // TODO: merge {onboarding,options}.{html,ts}.
 
-import { getPostUrl, setPostUrl } from "./util";
+import { setPostUrl } from "../../utils/util";
 
 const urlInput = document.getElementById("postUrl") as HTMLInputElement;
 const saveButton = document.getElementById("saveButton") as HTMLButtonElement;
@@ -13,13 +13,6 @@ function showStatus(message: string, type: string) {
 
 urlInput.addEventListener("input", () => showStatus("", ""));
 
-await (async function () {
-  const postUrl = await getPostUrl();
-  if (postUrl) {
-    urlInput.value = postUrl;
-  }
-})();
-
 saveButton.addEventListener("click", async () => {
   const postUrl = urlInput.value.trim();
   if (!postUrl) {
@@ -28,7 +21,10 @@ saveButton.addEventListener("click", async () => {
   }
   try {
     await setPostUrl(postUrl);
-    showStatus("Settings saved", "success");
+    showStatus("Settings saved, you may close this window.", "success");
+    // setTimeout(() => {
+    //  window.close();
+    // }, 3000);
   } catch (e) {
     const msg = `Failed: ${e instanceof Error ? e.message : e}`;
     showStatus(msg, "error");
