+++ packages/extension/src/entrypoints/options/index.html
@@ -4,6 +4,8 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Configuration - Empower Poster</title>
+  <!-- For wxt manifest.json auto-generation: -->
+  <meta name="manifest.open_in_tab" content="false" />
 </head>
 <body>
   <label for="postUrl">URL of your web API for JSON POST requests.</label>
