--- packages/extension/vitest.config.ts
@@ -1,8 +1,0 @@
-import { defineConfig } from 'vitest/config';
-
-export default defineConfig({
-  test: {
-    // needed? environment: 'node',
-    root: '.', // Ensures tests run from package root
-  },
-});
